Restructure karmabot
