+++ src/bot/cogs/__init__.py
@@ -1,0 +1,2 @@
+async def setup(bot):
+    ...
